+++ ext/stub/arrayaccessobj.zep.c
@@ -112,6 +112,7 @@ PHP_METHOD(Stub_ArrayAccessObj, offsetUnset)
 	zephir_fetch_params_without_memory_grow(1, 0, &offset);
 
 
+	zephir_unset_property_array(this_ptr, ZEND_STRL("test"), offset);
 	zephir_read_property(&_0, this_ptr, ZEND_STRL("test"), PH_NOISY_CC | PH_READONLY);
 	zephir_array_unset(&_0, offset, PH_SEPARATE);
 }