+++ ext/stub/assign.zep.h
@@ -92,9 +92,11 @@ PHP_METHOD(Stub_Assign, testArrayProperty);
 PHP_METHOD(Stub_Assign, testGlobalVarAssign);
 PHP_METHOD(Stub_Assign, testConstantKeyAssign);
 PHP_METHOD(Stub_Assign, testArrayBoolExpressionAssign);
+PHP_METHOD(Stub_Assign, testAssignBitwiseX);
 PHP_METHOD(Stub_Assign, testAssignSuperGlobals);
 PHP_METHOD(Stub_Assign, testAssignSuperGlobalsSERVER);
 PHP_METHOD(Stub_Assign, testAssignSuperGlobalsGET);
+PHP_METHOD(Stub_Assign, issue597);
 
 ZEND_BEGIN_ARG_INFO_EX(arginfo_stub_assign_gettestvar, 0, 0, 0)
 ZEND_END_ARG_INFO()
@@ -386,6 +388,11 @@ ZEND_END_ARG_INFO()
 ZEND_BEGIN_ARG_INFO_EX(arginfo_stub_assign_testarrayboolexpressionassign, 0, 0, 0)
 ZEND_END_ARG_INFO()
 
+ZEND_BEGIN_ARG_INFO_EX(arginfo_stub_assign_testassignbitwisex, 0, 0, 2)
+	ZEND_ARG_TYPE_INFO(0, a, IS_LONG, 0)
+	ZEND_ARG_TYPE_INFO(0, b, IS_LONG, 0)
+ZEND_END_ARG_INFO()
+
 ZEND_BEGIN_ARG_INFO_EX(arginfo_stub_assign_testassignsuperglobals, 0, 0, 0)
 ZEND_END_ARG_INFO()
 
@@ -393,6 +400,9 @@ ZEND_BEGIN_ARG_INFO_EX(arginfo_stub_assign_testassignsuperglobalsserver, 0, 0, 0
 ZEND_END_ARG_INFO()
 
 ZEND_BEGIN_ARG_INFO_EX(arginfo_stub_assign_testassignsuperglobalsget, 0, 0, 0)
+ZEND_END_ARG_INFO()
+
+ZEND_BEGIN_ARG_INFO_EX(arginfo_stub_assign_issue597, 0, 0, 0)
 ZEND_END_ARG_INFO()
 
 ZEPHIR_INIT_FUNCS(stub_assign_method_entry) {
@@ -609,6 +619,7 @@ ZEPHIR_INIT_FUNCS(stub_assign_method_entry) {
 #else
 	PHP_ME(Stub_Assign, testArrayBoolExpressionAssign, NULL, ZEND_ACC_PUBLIC)
 #endif
+	PHP_ME(Stub_Assign, testAssignBitwiseX, arginfo_stub_assign_testassignbitwisex, ZEND_ACC_PUBLIC)
 #if PHP_VERSION_ID >= 80000
 	PHP_ME(Stub_Assign, testAssignSuperGlobals, arginfo_stub_assign_testassignsuperglobals, ZEND_ACC_PUBLIC)
 #else
@@ -623,6 +634,11 @@ ZEPHIR_INIT_FUNCS(stub_assign_method_entry) {
 	PHP_ME(Stub_Assign, testAssignSuperGlobalsGET, arginfo_stub_assign_testassignsuperglobalsget, ZEND_ACC_PUBLIC)
 #else
 	PHP_ME(Stub_Assign, testAssignSuperGlobalsGET, NULL, ZEND_ACC_PUBLIC)
+#endif
+#if PHP_VERSION_ID >= 80000
+	PHP_ME(Stub_Assign, issue597, arginfo_stub_assign_issue597, ZEND_ACC_PUBLIC)
+#else
+	PHP_ME(Stub_Assign, issue597, NULL, ZEND_ACC_PUBLIC)
 #endif
 	PHP_FE_END
 };