+++ ext/stub/issue887.zep.c
@@ -1,0 +1,95 @@
+
+#ifdef HAVE_CONFIG_H
+#include "../ext_config.h"
+#endif
+
+#include <php.h>
+#include "../php_ext.h"
+#include "../ext.h"
+
+#include <Zend/zend_operators.h>
+#include <Zend/zend_exceptions.h>
+#include <Zend/zend_interfaces.h>
+
+#include "kernel/main.h"
+#include "kernel/memory.h"
+#include "kernel/object.h"
+#include "kernel/operators.h"
+
+
+/**
+ * @issue https://github.com/zephir-lang/zephir/issues/887
+ */
+ZEPHIR_INIT_CLASS(Stub_Issue887)
+{
+	ZEPHIR_REGISTER_CLASS(Stub, Issue887, stub, issue887, stub_issue887_method_entry, 0);
+
+	zend_declare_property_null(stub_issue887_ce, SL("data"), ZEND_ACC_PUBLIC);
+	return SUCCESS;
+}
+
+PHP_METHOD(Stub_Issue887, __construct)
+{
+	zval _0;
+	zephir_method_globals *ZEPHIR_METHOD_GLOBALS_PTR = NULL;
+	zval *this_ptr = getThis();
+
+	ZVAL_UNDEF(&_0);
+
+
+	ZEPHIR_MM_GROW();
+
+	ZEPHIR_INIT_VAR(&_0);
+	array_init(&_0);
+	zephir_update_property_zval(this_ptr, ZEND_STRL("data"), &_0);
+	ZEPHIR_MM_RESTORE();
+}
+
+PHP_METHOD(Stub_Issue887, add)
+{
+	zval *key, key_sub, *value, value_sub;
+	zval *this_ptr = getThis();
+
+	ZVAL_UNDEF(&key_sub);
+	ZVAL_UNDEF(&value_sub);
+#if PHP_VERSION_ID >= 80000
+	bool is_null_true = 1;
+	ZEND_PARSE_PARAMETERS_START(2, 2)
+		Z_PARAM_ZVAL(key)
+		Z_PARAM_ZVAL(value)
+	ZEND_PARSE_PARAMETERS_END();
+#endif
+
+
+	zephir_fetch_params_without_memory_grow(2, 0, &key, &value);
+
+
+	zephir_update_property_array(this_ptr, SL("data"), key, value);
+}
+
+PHP_METHOD(Stub_Issue887, fetch_array)
+{
+	zval *this_ptr = getThis();
+
+
+
+	RETURN_MEMBER(getThis(), "data");
+}
+
+PHP_METHOD(Stub_Issue887, fetch_object)
+{
+	zval _0;
+	zephir_method_globals *ZEPHIR_METHOD_GLOBALS_PTR = NULL;
+	zval *this_ptr = getThis();
+
+	ZVAL_UNDEF(&_0);
+
+
+	ZEPHIR_MM_GROW();
+
+	ZEPHIR_OBS_VAR(&_0);
+	zephir_read_property(&_0, this_ptr, ZEND_STRL("data"), PH_NOISY_CC);
+	zephir_convert_to_object(&_0);
+	RETURN_CCTOR(&_0);
+}
+