+++ tests/Extension/ArrayAccessTest.php
@@ -17,12 +17,37 @@ use PHPUnit\Framework\TestCase;
 
 final class ArrayAccessTest extends TestCase
 {
+    private array $defaultUnsetData = [
+        'key_a' => 'marcin',
+        'key_b' => 'paula',
+        3 => 'long value',
+        //3.14 => 'double value', // Not supported yet
+        //false => 'bool value', // Not supported yet
+        //null => 'null value', // Not supported yet
+    ];
+
     public function testTest(): void
     {
         $class = new \Stub\ArrayAccessTest();
 
         $this->assertTrue($class->exits('one'));
         $this->assertSame(2, $class->get());
+    }
+
+    public function testUnsetByKeyFromArray(): void
+    {
+        $class = new \Stub\ArrayAccessTest();
+
+        $this->assertSame([], $class->unsetByKeyFromArray('not-exist', []));
+    }
+
+    public function testUnsetByKeyFromProperty(): void
+    {
+        $class = new \Stub\ArrayAccessTest();
+
+        $this->assertSame([], $class->unsetByKeyFromProperty('ok', ['ok' => true]));
+        $this->assertSame(['another' => 'value'], $class->unsetByKeyFromProperty('ok', ['ok' => true, 'another' => 'value']));
+        $this->assertSame([], $class->unsetByKeyFromProperty('not-exist', []));
     }
 
     /**
@@ -43,5 +68,89 @@ final class ArrayAccessTest extends TestCase
         $class = new \Stub\ArrayAccessTest();
 
         $this->assertFalse($class->issue1155());
+    }
+
+    public function testIssue1094(): void
+    {
+        $class = new \Stub\ArrayAccessTest();
+
+        $this->assertFalse($class->issue1094Test1());
+        $this->assertFalse($class->issue1094Test1([]));
+        $this->assertFalse($class->issue1094Test1(['test' => 'ok']));
+
+        $this->assertTrue($class->issue1094Test2());
+        $this->assertFalse($class->issue1094Test2([]));
+        $this->assertFalse($class->issue1094Test2(['test' => 'ok']));
+
+        $this->assertTrue($class->issue1094Test3());
+        $this->assertFalse($class->issue1094Test3([]));
+        $this->assertFalse($class->issue1094Test3(['test' => 'ok']));
+    }
+
+    /**
+     * @issue https://github.com/zephir-lang/zephir/issues/1086
+     */
+    public function testIssue1086StaticallyCalledFunctionWithArrayAsArgMustReturnArray(): void
+    {
+        $class = new \Stub\ArrayAccessTest();
+
+        $actual = $class->issue1086WontNullArrayAfterPassViaStaticWithStrictParams();
+        $this->assertSame(['test' => 123], $actual);
+
+        $actual = $class->issue1086WontNullArrayAfterPassViaStaticWithoutStrictParams();
+        $this->assertSame(['test' => 123], $actual);
+    }
+
+    /**
+     * @issue https://github.com/zephir-lang/zephir/issues/1259
+     */
+    public function testIssue1259CheckUnsetKeyFromArray(): void
+    {
+        $class = new \Stub\ArrayAccessTest();
+
+        $expected = [
+            ['key_a' => 'marcin', 'key_b' => 'paula'],
+            ['key_b' => 'paula'],
+        ];
+
+        $this->assertSame($expected, $class->issue1259UnsetKeyFromArrayInternalVariable());
+    }
+
+    /**
+     * @issue https://github.com/zephir-lang/zephir/issues/1259
+     */
+    public function testIssue1259CheckUnsetStringKeyFromArrayProperty(): void
+    {
+        $class = new \Stub\ArrayAccessTest();
+
+        $this->assertSame(
+            [
+                $this->defaultUnsetData,
+                [
+                    'key_b' => 'paula',
+                    3 => 'long value',
+                ],
+            ],
+            $class->issue1259UnsetStringKeyFromArrayProperty()
+        );
+    }
+
+    /**
+     * @issue https://github.com/zephir-lang/zephir/issues/1259
+     */
+    public function testIssue1259CheckUnsetLongKeyFromArrayProperty(): void
+    {
+        $class = new \Stub\ArrayAccessTest();
+
+        $this->assertSame(
+            [
+                $this->defaultUnsetData,
+                [
+                    'key_a' => 'marcin',
+                    'key_b' => 'paula',
+                ],
+            ],
+            $class->issue1259UnsetLongKeyFromArrayProperty()
+        );
     }
 }